Change some stuff

+++ events/message.js
@@ -1,9 +1,9 @@
 exports.run = (client, msg) => {
 	if (msg.author.bot) return;
 	if (msg.channel.type !== 'text') return msg.reply('You must run the commands on a Discord server on which the Discord Bot is available');
-	const prefix = '-';
-	if (!msg.content.startsWith(prefix)) return;
-	var command = msg.content.split(' ')[0].slice(prefix.length).toLowerCase();
+	const tableload = client.guildconfs.get(msg.guild.id);
+	if (!msg.content.startsWith(tableload.prefix)) return;
+	var command = msg.content.split(' ')[0].slice(tableload.prefix.length).toLowerCase();
 	var args = msg.content.split(' ').slice(1);
 	let cmd;
 	if (client.commands.has(command)) {
@@ -13,8 +13,8 @@ exports.run = (client, msg) => {
 	}
 	if (cmd) {
 		cmd.run(client, msg, args);
-		// if (tableload.commanddel === 'true') {
+		if (tableload.commanddel === 'true') {
 			msg.delete();
-		// }
+		}
 	}
 };
